move apacheKdcForSpnego class into spnego_fat
